Updates to card duplication  - Duplicates into My Collection at the end, and sends you there, triggering an AlertModal  - Includes minor refactor of AlertModals

+++ app/javascript/ui/roles/RoleSelect.js
@@ -45,10 +45,10 @@ class RoleSelect extends React.Component {
         ${this.renderName()} from this group?`
       confirmText = 'Remove'
     }
-    uiStore.openAlertModal({
+    uiStore.confirm({
       prompt,
       confirmText,
-      iconName: 'LeaveIcon',
+      iconName: 'Leave',
       onConfirm: () => this.deleteRole(true),
     })
   }
